pyproject.toml
[tool.poetry]
name = "duckdb-processing-persistence"
version = "0.1.0"
description = ""
authors = ["Ant4res <pr1v4t3.lm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
duckdb = "^0.10.1"
jupyter = "^1.0.0"
google-auth = "^2.28.2"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.122.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
